move tracing to infra package
